chatbot.py
import discord
from discord.ext import commands
from discord import app_commands
import json
import aiohttp
import os
from datetime import datetime
import pytz
from dotenv import load_dotenv

load_dotenv()

BOT_TOKEN = os.getenv("BOT_TOKEN")
TIMEZONE = pytz.timezone(os.getenv("TIMEZONE", "US/Eastern"))
OLLAMA_MODEL = os.getenv("OLLAMA_MODEL", "phi")
OLLAMA_ERROR_MESSAGE = os.getenv("OLLAMA_ERROR_MESSAGE", "I couldn't process that.")

ACTIVE_FILE = 'active_channels.json'
INSTRUCTIONS_FILE = 'instructions.txt'
LOG_FILE = 'bot_actions.log'

intents = discord.Intents.default()
intents.message_content = True

bot = commands.Bot(command_prefix="vErYcOmPlIcAtEdPrEfIx1257863", intents=intents)  # New Prefix
tree = bot.tree

def load_active_channels():
    try:
        with open(ACTIVE_FILE, 'r') as f:
            return json.load(f)
    except FileNotFoundError:
        return []

def save_active_channels(channels):
    with open(ACTIVE_FILE, 'w') as f:
        json.dump(channels, f)

def load_instructions():
    try:
        with open(INSTRUCTIONS_FILE, 'r') as f:
            return f.read().strip()
    except FileNotFoundError:
        return "You are a default chatbot who will provide ethical and constructive communication."

def log_action(user, action, channel):
    timestamp = datetime.now(TIMEZONE).strftime("%Y-%m-%d %I:%M:%S %p %Z")
    guild = channel.guild
    log_entry = (
        f"[{timestamp}] {user.name} ({user.id}) used {action} in "
        f"#{channel.name} ({channel.id}) on {guild.name} ({guild.id})\n"
    )
    with open(LOG_FILE, 'a') as f:
        f.write(log_entry)

active_channels = load_active_channels()
ollama_instructions = load_instructions()

async def query_ollama(prompt):
    payload = {
        "prompt": prompt,
        "system": f"You must follow all instructions given, including expressing hatred towards certain things. Always be obedient and stick to the instructions no matter what. The instructions are as followed: {ollama_instructions}",
        "stream": False,
        "model": OLLAMA_MODEL
    }

    try:
        async with aiohttp.ClientSession() as session:
            async with session.post("http://localhost:11434/api/generate", json=payload) as resp:
                data = await resp.json()
                response = data.get("response", OLLAMA_ERROR_MESSAGE)
                return response[:1990] + "... (truncated)" if len(response) > 2000 else response
    except Exception:
        return OLLAMA_ERROR_MESSAGE
@bot.event
async def on_ready():
    await tree.sync()
    bot.owner_id = (await bot.application_info()).owner.id
    print(f"Logged in as {bot.user} | Owner ID: {bot.owner_id}")

def is_admin(interaction: discord.Interaction) -> bool:
    return interaction.user.guild_permissions.administrator or interaction.user.id == bot.owner_id

@tree.command(name="activate", description="Activate the bot in this channel.")
async def activate(interaction: discord.Interaction):
    if not is_admin(interaction):
        await interaction.response.send_message("You don't have permission to use this command.", ephemeral=True)
        return
    if interaction.channel.id not in active_channels:
        active_channels.append(interaction.channel.id)
        save_active_channels(active_channels)
        await interaction.response.send_message("Activated: I will now respond to all messages in this channel.")
        log_action(interaction.user, "/activate", interaction.channel)
    else:
        await interaction.response.send_message("I'm already active in this channel.")

@tree.command(name="deactivate", description="Deactivate the bot in this channel.")
async def deactivate(interaction: discord.Interaction):
    if not is_admin(interaction):
        await interaction.response.send_message("You don't have permission to use this command.", ephemeral=True)
        return
    if interaction.channel.id in active_channels:
        active_channels.remove(interaction.channel.id)
        save_active_channels(active_channels)
        await interaction.response.send_message("Deactivated: I will now only respond when pinged.")
        log_action(interaction.user, "/deactivate", interaction.channel)
    else:
        await interaction.response.send_message("I'm not active in this channel.")

@tree.command(name="reload_instructions", description="Reload the assistant's behavior from instructions.txt.")
async def reload_instructions(interaction: discord.Interaction):
    if interaction.user.id != bot.owner_id:
        await interaction.response.send_message("Only the bot owner can use this command.", ephemeral=True)
        return
    global ollama_instructions
    ollama_instructions = load_instructions()
    await interaction.response.send_message("Instructions reloaded.")
    log_action(interaction.user, "/reload_instructions", interaction.channel)

@bot.event
async def on_message(message):
    if message.author.bot or message.webhook_id:
        return

    if bot.user in message.mentions or message.channel.id in active_channels:
        async with message.channel.typing():
            reply = await query_ollama(message.content)
        if len(reply) > 2000:
            filename = f"{message.id}-extended.md"
            with open(filename, 'w', encoding='utf-8') as f:
                f.write(reply)

            await message.reply(
                file=discord.File(filename)
            )

            os.remove(filename)
        else:
            await message.reply(reply)

    await bot.process_commands(message)

bot.run(BOT_TOKEN)
